Default scroll direction Down
There is an edge case where if your step element is visible when `scrollTop` is 0, the scroll `direction` will be `up`. I.e. when you first load your page, the first scroll direction you'll see in `onStepEnter` will be `up`.

While admittedly an edge case, I think we should switch the logic so that the default scroll direction is `down`